Add TC type to temperature example.
Why
---

Setting the TC type for a temperature Phidgets isn't as easy as one may
think. Phidgets recommends setting the TC type from the on_attach
handler.

This change adds a TC type into the CLI args for the temperature
example, and shows how to set the TC type from the on_attach handler.
This is a bit more tricky with Rust vs. other languages.

How
---

- Add `-t` CLI switch that takes a single character
- Set the TC type after attachment, defaulting to K